+++ lib/async.dart
@@ -22,6 +22,7 @@ export 'src/delegate/stream_sink.dart';
 export 'src/delegate/stream_subscription.dart';
 export 'src/future_group.dart';
 export 'src/lazy_stream.dart';
+export 'src/notifier.dart';
 export 'src/null_stream_sink.dart';
 export 'src/restartable_timer.dart';
 export 'src/result/error.dart';